Change path for projects images
In order to fix D'arc's image loading,
we change the path where the images are stored.
Instead of using the name of the project (i.e. "D'Arc"),
we use the slug of the project (i.e. "darc").
This way, we assure that are no special caracters
in the file path (i.e. ')